requirements.txt
annotated-types==0.6.0
anyio==4.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
gigachat==0.1.8
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.25.1
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
requests==2.31.0
sniffio==1.3.0
SpeechRecognition==3.10.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
